+++ parsing/src/main/java/org/apache/calcite/sql/parsing/Parsing.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.calcite.sql.bigquery;
+package org.apache.calcite.sql.parsing;
 
 /** Parsing folder. */
 @SuppressWarnings("unused")
